i killed taedeo (#2321)
